+++ src/main/java/org/opensearch/flowframework/workflow/NoOpStep.java
@@ -9,8 +9,17 @@
 package org.opensearch.flowframework.workflow;
 
 import org.opensearch.action.support.PlainActionFuture;
+import org.opensearch.common.unit.TimeValue;
+import org.opensearch.common.util.concurrent.FutureUtils;
+import org.opensearch.core.rest.RestStatus;
+import org.opensearch.flowframework.exception.WorkflowStepException;
+import org.opensearch.flowframework.util.ParseUtils;
 
+import java.util.Collections;
 import java.util.Map;
+import java.util.Set;
+
+import static org.opensearch.flowframework.common.CommonValue.DELAY_FIELD;
 
 /**
  * A workflow step that does nothing. May be used for synchronizing other actions.
@@ -32,6 +41,29 @@ public class NoOpStep implements WorkflowStep {
         Map<String, String> params
     ) {
         PlainActionFuture<WorkflowData> future = PlainActionFuture.newFuture();
+
+        Set<String> requiredKeys = Collections.emptySet();
+        Set<String> optionalKeys = Set.of(DELAY_FIELD);
+
+        try {
+            Map<String, Object> inputs = ParseUtils.getInputsFromPreviousSteps(
+                requiredKeys,
+                optionalKeys,
+                currentNodeInputs,
+                outputs,
+                previousNodeInputs,
+                params
+            );
+            if (inputs.containsKey(DELAY_FIELD)) {
+                long delay = TimeValue.parseTimeValue(inputs.get(DELAY_FIELD).toString(), DELAY_FIELD).millis();
+                Thread.sleep(delay);
+            }
+        } catch (IllegalArgumentException iae) {
+            throw new WorkflowStepException(iae.getMessage(), RestStatus.BAD_REQUEST);
+        } catch (InterruptedException e) {
+            FutureUtils.cancel(future);
+        }
+
         future.onResponse(WorkflowData.EMPTY);
         return future;
     }