+++ src/test/java/org/opensearch/flowframework/workflow/NoOpStepTests.java
@@ -9,10 +9,14 @@
 package org.opensearch.flowframework.workflow;
 
 import org.opensearch.action.support.PlainActionFuture;
+import org.opensearch.flowframework.exception.WorkflowStepException;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.io.IOException;
 import java.util.Collections;
+import java.util.Map;
+
+import static org.opensearch.flowframework.common.CommonValue.DELAY_FIELD;
 
 public class NoOpStepTests extends OpenSearchTestCase {
 
@@ -27,5 +31,33 @@ public class NoOpStepTests extends OpenSearchTestCase {
             Collections.emptyMap()
         );
         assertTrue(future.isDone());
+    }
+
+    public void testNoOpStepDelay() throws IOException, InterruptedException {
+        NoOpStep noopStep = new NoOpStep();
+        WorkflowData delayData = new WorkflowData(Map.of(DELAY_FIELD, "1s"), null, null);
+
+        long start = System.nanoTime();
+        PlainActionFuture<WorkflowData> future = noopStep.execute(
+            "nodeId",
+            delayData,
+            Collections.emptyMap(),
+            Collections.emptyMap(),
+            Collections.emptyMap()
+        );
+        assertTrue(future.isDone());
+        // Sleep isn't exactly accurate so leave 100ms of roundoff
+        assertTrue(System.nanoTime() - start > 900_000_000L);
+    }
+
+    public void testNoOpStepParse() throws IOException {
+        NoOpStep noopStep = new NoOpStep();
+        WorkflowData delayData = new WorkflowData(Map.of(DELAY_FIELD, "foo"), null, null);
+
+        Exception ex = assertThrows(
+            WorkflowStepException.class,
+            () -> noopStep.execute("nodeId", delayData, Collections.emptyMap(), Collections.emptyMap(), Collections.emptyMap())
+        );
+        assertEquals("failed to parse setting [delay] with value [foo] as a time value: unit is missing or unrecognized", ex.getMessage());
     }
 }